feat: adapt to LeviLamina 26.20.0

+++ src/LOICollectionA/utils/core/MathUtils.h
@@ -1,5 +1,5 @@
 #pragma once
 
 namespace MathUtils {
-    int pow(int base, int exponent);
+    double pow(double base, double exponent);
 }
